Oferta relampago

+++ src/main/java/ecommerce/eco/model/request/ReviewRequest.java
@@ -14,7 +14,7 @@ import java.time.ZonedDateTime;
 public class ReviewRequest {
     @NotEmpty(message = "idProduct no puede estar vacio")
     private Long idProduct;
-    private int score; // estrellas o puntaje
+    private double score; // estrellas o puntaje
     private String username; // nombre del usuario
     private String comment; // comentario
     //@JsonFormat(pattern="yyyy-MM-dd HH:mm:ss")
